CPatternDialog運作正常，檢查目前改過的所有程式碼，正常
修正 D21最後一點為中心點沒有優先量測
修正 CrsTlk刪掉一點，刪掉相同顏色&點位，BkStatus沒有差異，造成失去辨識功能
整理程式碼（並且將可以加上const 的部份加上const）

Bullet、Cartridge2、ColorRef、DNA、Nucleotide、RNA、TranScripter、BkCrossTalk、BkMaker、BkNits、Ca210、Ca210real、Ca210sim、CFileDlg、CCaSetupDlg、CEnterValueDlg、CMsrItemDlg、CPatternDlg、ValuesLabel、Circle
測試模擬儀器量測一輪
測試真實儀器量測一輪

+++ BkNits.cpp
@@ -13,7 +13,7 @@ void BkNits::n2cPos(ColorRef& Bkclr)
         //量測抓值
         if (m_pCa210->Measure() == CA_ZeroCalMode)
             AfxMessageBox("檔位不在MEAS");
-        fLv = (m_pCa210->GetMsrData().oFlt(VluK_Lv))*100;  //m_IProbe.GetLv();
+        fLv = (int)((m_pCa210->GetMsrData().oFlt(VluK_Lv))*100);  //m_IProbe.GetLv();
     }
 	
     while(fLv < m_Nits)   //若亮度還在5以下，就...變亮
@@ -25,7 +25,7 @@ void BkNits::n2cPos(ColorRef& Bkclr)
         Sleep(60);
 		//      量測抓值
         m_pCa210->Measure();
-        fLv = m_pCa210->GetMsrData().oFlt(VluK_Lv)*100;  //m_IProbe.GetLv();
+        fLv = (int)((m_pCa210->GetMsrData().oFlt(VluK_Lv))*100);  //m_IProbe.GetLv();
     }
 	_SetBkColor(Bkclr);
 }
@@ -45,7 +45,7 @@ void BkNits::n2cNeg(ColorRef& Bkclr)
 		//        量測抓值
         if (m_pCa210->Measure() == CA_ZeroCalMode)
             AfxMessageBox("檔位不在MEAS");
-        fLv = m_pCa210->GetMsrData().oFlt(VluK_Lv)*100;  //m_IProbe.GetLv();
+        fLv = (int)((m_pCa210->GetMsrData().oFlt(VluK_Lv))*100);  //m_IProbe.GetLv();
     }
     
     while(fLv > m_Nits)   //若亮度還在5以下，就...變亮
@@ -57,7 +57,7 @@ void BkNits::n2cNeg(ColorRef& Bkclr)
 		Sleep(60);
         //量測抓值
         m_pCa210->Measure();
-        fLv = m_pCa210->GetMsrData().oFlt(VluK_Lv)*100;  //m_IProbe.GetLv();
+        fLv = (int)((m_pCa210->GetMsrData().oFlt(VluK_Lv))*100);  //m_IProbe.GetLv();
     }
 	_SetBkColor(Bkclr);
 }
